3.11 Se quita la clase MainActivity
