+++ ADictionaryEcho/app/src/main/java/model/ModelModule.java
@@ -1,7 +1,5 @@
 package model;
 
-import android.arch.persistence.room.Database;
-
 import com.google.gson.Gson;
 
 import model.service.TranslatorServiceImpl;
@@ -13,7 +11,7 @@ public class ModelModule {
     private TranslatorModel dicModel;
 
     private ModelModule() {
-        dicModel =  new TranslatorModelConcrete2(new TranslatorServiceImpl(new YandexApiConnection().getYandex(),new Gson()), DataBase.getInstance());
+        dicModel =  new TranslatorModelConcrete(new TranslatorServiceImpl(new YandexApiConnection().getYandex(),new Gson()), DataBase.getInstance());
     }
 
     public static ModelModule getInstance() {
